refactor: use hooks and modularize router
