+++ .gitignore
@@ -25,3 +25,7 @@ dist-ssr
 
 # Environment variables
 .env
+
+
+# TypeScript
+*.tsbuildinfo
