+++ src/hooks/usePoints.ts
@@ -1,0 +1,32 @@
+import { useCallback, useEffect } from "react";
+
+import { useGlobalState } from "../contexts/global";
+import { isUserRegistered } from "../utils/helpers";
+
+export const usePoints = () => {
+  const { username, setPoints } = useGlobalState();
+
+  const getPoints = useCallback(async () => {
+    try {
+      if (isUserRegistered()) {
+        fetch(process.env.BACKEND_API_URL + "/points/" + username).then((resp) =>
+          resp.text().then((data) => {
+            setPoints(Number(data));
+          })
+        );
+      }
+    } catch (error) {
+      console.error("error fetching points: ", error);
+    }
+  }, [username, setPoints]);
+
+  useEffect(() => {
+    getPoints();
+    const interval = setInterval(() => {
+      getPoints();
+    }, 1000 * 60);
+    return () => clearInterval(interval);
+  }, [getPoints]);
+
+  return { getPoints };
+};
