+++ src/router.tsx
@@ -1,13 +1,17 @@
-import { PrivateKey } from "@ethersphere/bee-js";
-import { MessageData } from "@solarpunkltd/comment-system";
-import { loadLatestComments } from "@solarpunkltd/comment-system-ui";
-import { Wallet } from "ethers";
-import { ReactElement, useCallback, useEffect, useState } from "react";
+import { ReactElement, useEffect } from "react";
 import { Route, Routes, useLocation, useNavigate } from "react-router-dom";
 
 import Gamification from "./components/Gamification/Gamification";
-import { NoteItemProps } from "./components/NoteItem/NoteItem";
 import { useGlobalState } from "./contexts/global";
+import { useBeePing } from "./hooks/useBeePing";
+import { useGamification } from "./hooks/useGamification";
+import { useNotes } from "./hooks/useNotes";
+import { usePoints } from "./hooks/usePoints";
+import { usePreloadTalks } from "./hooks/usePreloadTalks";
+import { usePrevLocation } from "./hooks/usePrevLocation";
+import { useRouteProtection } from "./hooks/useRouteProtection";
+import { useSessionData } from "./hooks/useSessionData";
+import { useViewportHeight } from "./hooks/useViewportHeight";
 import Agenda from "./pages/Agenda/Agenda";
 import ClaimRewardPage from "./pages/ClaimReward/ClaimRevard";
 import ContentFilter from "./pages/ContentFilter/ContentFilter";
@@ -27,352 +31,44 @@ import Welcome1 from "./pages/Welcome1/Welcome1";
 import Welcome2 from "./pages/Welcome2/Welcome2";
 import Welcome3 from "./pages/Welcome3/Welcome3";
 import Welcome4 from "./pages/Welcome4/Welcome4";
-import { Session } from "./types/session";
-import { TalkComments } from "./types/talkComment";
-import { getFeedUpdate, getTopic } from "./utils/bee";
-import {
-  ADDRESS_HEX_LENGTH,
-  CATEGORIES,
-  FIVE_MINUTES,
-  MAX_COMMENTS_LOADED,
-  MAX_SESSIONS_SHOWN,
-  ROUTES,
-  SELF_NOTE_TOPIC,
-  SPACES_KEY,
-} from "./utils/constants";
-import { findSlotStartIx, getLocalPrivateKey, getSessionsByDay, isUserRegistered } from "./utils/helpers";
+import { ROUTES } from "./utils/constants";
 
 const MainRouter = (): ReactElement => {
-  const {
-    showGamification,
-    setShowGamification,
-    points,
-    setPoints,
-    username,
-    sessions,
-    setSessions,
-    recentSessions,
-    setRecentSessions,
-    loadedTalks,
-    setLoadedTalks,
-    notes,
-    setNotes,
-    setTalkActivity,
-    setSpacesActivity,
-  } = useGlobalState();
-  const [isBeeRunning, setBeeRunning] = useState<boolean>(false);
-  const [recentSessionIx, setRecentSessionIx] = useState<number>(0);
-  const [time, setTime] = useState<number>(new Date().getTime());
-  const [noteRawTopics, setNoteRawTopics] = useState<string[]>([]);
+  const { showGamification, sessions, points } = useGlobalState();
   const location = useLocation();
   const navigate = useNavigate();
-  const [prevLocation, setPrevLocation] = useState<string | null>(null);
-  const [isFirstRender, setIsFirstRender] = useState<boolean>(true);
 
-  const setVhVariable = () => {
-    const vh = window.innerHeight * 0.01;
-    document.documentElement.style.setProperty("--vh", `${vh}px`);
-  };
+  const { isBeeRunning } = useBeePing();
+
+  const { filterRecentSessions } = useSessionData(isBeeRunning);
 
   useEffect(() => {
-    return () => {
-      setPrevLocation(location.pathname);
-    };
-  }, [location]);
-
-  useEffect(() => {
-    setVhVariable();
-    window.addEventListener("resize", setVhVariable);
-    return () => {
-      window.removeEventListener("resize", setVhVariable);
-    };
-  }, []);
-
-  const checkBee = async () => {
-    try {
-      await fetch(process.env.BEE_API_URL + "/bytes/" + process.env.HEALTH_CHECK_DATA_REF);
-      if (!isBeeRunning) {
-        setBeeRunning(true);
-        console.debug("Bee is running");
-      }
-    } catch (error) {
-      if (isBeeRunning) {
-        setBeeRunning(false);
-        console.error("Bee stopped running, error: ", error);
-      }
+    if (sessions) {
+      filterRecentSessions(sessions);
     }
-  };
+  }, [sessions, filterRecentSessions]);
 
-  const getPoints = async () => {
-    try {
-      if (isUserRegistered()) {
-        fetch(process.env.BACKEND_API_URL + "/points/" + username).then((resp) =>
-          resp.text().then((data) => {
-            setPoints(Number(data));
-          })
-        );
-      }
-    } catch (error) {
-      console.error("error fetching points: ", error);
-    }
-  };
+  const { preLoadTalks, calcActivity, calcSpacesActivity } = usePreloadTalks();
 
   useEffect(() => {
-    checkBee();
-    const interval = setInterval(() => {
-      checkBee();
-    }, FIVE_MINUTES);
-    return () => clearInterval(interval);
-  }, []);
-
-  useEffect(() => {
-    getPoints();
-    const interval = setInterval(() => {
-      getPoints();
-    }, 1000 * 60);
-    return () => clearInterval(interval);
-  }, [username]);
-
-  const fetchFeedUpdate = useCallback(async () => {
-    if (isBeeRunning) {
-      const rawFeedTopicSession = "sessions";
-      const dataStr = await getFeedUpdate(process.env.FEED_OWNER_ADDRESS as string, rawFeedTopicSession);
-      const data = new Map<string, Session[]>(Object.entries(JSON.parse(dataStr)));
-
-      const spacesSessions: Session[] = [];
-      for (let i = 0; i < CATEGORIES.length; i++) {
-        const cat = CATEGORIES[i];
-        spacesSessions.push({
-          id: cat,
-          sourceId: cat,
-          title: cat,
-          track: cat,
-          slot_start: new Date().toLocaleString(),
-          slot_end: new Date().toLocaleString(),
-          slot_roomId: cat,
-        });
-      }
-
-      data.set(SPACES_KEY, spacesSessions);
-      if (data.size !== 0) {
-        console.debug("session data updated");
-        setSessions(() => data);
-      } else {
-        console.debug("session data empty");
-      }
+    if (isBeeRunning && sessions && sessions.size > 0) {
+      preLoadTalks();
+      calcActivity();
+      calcSpacesActivity();
     }
-  }, [isBeeRunning]);
+  }, [isBeeRunning, sessions, preLoadTalks, calcActivity, calcSpacesActivity]);
 
-  useEffect(() => {
-    fetchFeedUpdate();
-    const interval = setInterval(async () => {
-      fetchFeedUpdate();
-    }, FIVE_MINUTES);
+  const { prevLocation } = usePrevLocation(location);
 
-    return () => clearInterval(interval);
-  }, [fetchFeedUpdate]);
+  useViewportHeight();
 
-  useEffect(() => {
-    if (isFirstRender) {
-      setIsFirstRender(false);
-    } else {
-      if (points === 1 || points === 5 || points === 10) {
-        setShowGamification(true);
-      }
-    }
-  }, [points]);
+  usePoints();
 
-  // the uploaded sessions are already sorted by time and day
-  // find the first session that starts after the current time
-  const filterRecentSessions = () => {
-    const day = "all";
-    const sessionsByDay = getSessionsByDay(sessions, day);
-    if (sessionsByDay.length != 0) {
-      const mostRecentSessions = new Array<Session>(MAX_SESSIONS_SHOWN);
-      let firstSessionIx = findSlotStartIx(recentSessionIx, sessionsByDay, time);
-      firstSessionIx = firstSessionIx > MAX_SESSIONS_SHOWN - 1 ? firstSessionIx : MAX_SESSIONS_SHOWN - 1;
+  useNotes();
 
-      for (let i = 0; i < MAX_SESSIONS_SHOWN && 0 < sessionsByDay.length - firstSessionIx - i; i++) {
-        const recentIx = firstSessionIx - i;
-        mostRecentSessions[i] = sessionsByDay[recentIx];
-      }
-      setRecentSessionIx(firstSessionIx);
-      setRecentSessions(mostRecentSessions);
-    }
-  };
+  useGamification();
 
-  useEffect(() => {
-    const interval = setInterval(async () => {
-      setTime(new Date().getTime());
-    }, FIVE_MINUTES);
-
-    return () => clearInterval(interval);
-  }, []);
-
-  useEffect(() => {
-    filterRecentSessions();
-  }, [sessions, time]);
-
-  const preLoadTalks = async () => {
-    try {
-      const preLoadedTalks: TalkComments[] = [];
-      const commentPromises: Promise<{ talkId: string; messages: any[] }>[] = [];
-      const talkIds: string[] = [];
-      for (let i = 0; i < recentSessions.length; i++) {
-        const sessionId = recentSessions[i].id;
-        const rawTalkTopic = getTopic(sessionId);
-        const signer = new PrivateKey(rawTalkTopic);
-        // only load the talks that are not already loaded
-        if (loadedTalks) {
-          // talkids include a "version" suffix
-          const foundIx = loadedTalks.findIndex((talk) => talk.talkId.includes(rawTalkTopic));
-          if (foundIx > -1) {
-            preLoadedTalks.push(loadedTalks[foundIx]);
-            continue;
-          }
-        }
-        commentPromises.push(loadLatestComments(rawTalkTopic, signer.publicKey().address.toString(), process.env.BEE_API_URL, MAX_COMMENTS_LOADED));
-        talkIds.push(rawTalkTopic);
-      }
-
-      await Promise.allSettled(commentPromises).then((results) => {
-        results.forEach((result, i) => {
-          if (result.status === "fulfilled") {
-            preLoadedTalks.push({
-              talkId: talkIds[i],
-              messages: result.value.messages,
-            });
-          } else {
-            console.debug(`preloading talks error: `, result.reason);
-          }
-        });
-      });
-
-      setLoadedTalks(preLoadedTalks.length > 0 ? preLoadedTalks : undefined);
-    } catch (error) {
-      console.debug("preloading talks error: ", error);
-    }
-  };
-
-  useEffect(() => {
-    preLoadTalks();
-  }, [recentSessions]);
-
-  const calcActivity = async () => {
-    if (loadedTalks) {
-      const tmpActiveVisitors = new Map<string, number>();
-      for (let i = 0; i < recentSessions.length; i++) {
-        const foundIx = loadedTalks.findIndex((talk) => talk.talkId.includes(recentSessions[i].id));
-        if (foundIx > -1) {
-          tmpActiveVisitors.set(recentSessions[i].id, loadedTalks[foundIx].messages?.length ?? 0);
-        }
-      }
-      setTalkActivity(tmpActiveVisitors);
-    }
-  };
-
-  const calcSpacesActivity = async () => {
-    const spacesSessions = getSessionsByDay(sessions, SPACES_KEY);
-    const spacesPromises: Promise<{ talkId: string; messages: MessageData[] }>[] = [];
-    try {
-      for (let i = 0; i < spacesSessions.length; i++) {
-        const rawTalkTopic = getTopic(spacesSessions[i].id);
-        const signer = new PrivateKey(rawTalkTopic);
-        spacesPromises.push(loadLatestComments(rawTalkTopic, signer.publicKey().address.toString(), process.env.BEE_API_URL, MAX_COMMENTS_LOADED));
-      }
-
-      const tmpActivity = new Map<string, number>();
-      await Promise.allSettled(spacesPromises).then((results) => {
-        results.forEach((result, i) => {
-          if (result.status === "fulfilled") {
-            tmpActivity.set(spacesSessions[i].id, result.value.messages.length);
-          } else {
-            console.error(`fetching user count of talks error: `, result.reason);
-          }
-        });
-      });
-      setSpacesActivity(tmpActivity);
-    } catch (error) {
-      console.error("fetching user count of talks error: ", error);
-    }
-  };
-
-  useEffect(() => {
-    calcActivity();
-  }, [loadedTalks]);
-
-  useEffect(() => {
-    calcSpacesActivity();
-  }, [recentSessions]);
-
-  const fetchNotes = async () => {
-    const privKey = getLocalPrivateKey();
-    if (!privKey) {
-      console.error("private key not found - cannot fetch notes");
-      return;
-    }
-
-    const wallet = new Wallet(privKey);
-    const feedPromises: Promise<string>[] = [];
-    for (let i = 0; i < noteRawTopics.length; i++) {
-      const rawTopic = noteRawTopics[i];
-      feedPromises.push(getFeedUpdate(wallet.address, rawTopic));
-    }
-
-    const notesArray: string[] = [];
-    await Promise.allSettled(feedPromises).then((results) => {
-      results.forEach((result) => {
-        if (result.status === "fulfilled") {
-          notesArray.push(result.value);
-        } else {
-          console.error(`fetching note data error: `, result.reason);
-        }
-      });
-    });
-
-    const tmpNotes: NoteItemProps[] = [...notes];
-    for (let i = 0; i < notesArray.length; i++) {
-      let note: NoteItemProps | undefined = undefined;
-      try {
-        note = JSON.parse(notesArray[i]) as NoteItemProps;
-      } catch (error) {
-        console.error(`error parsing notes[${i}]:\n ${error}`);
-        continue;
-      }
-      const found = notes.find((n) => n.id === note.id);
-      if (!found && note !== undefined) {
-        tmpNotes.push(note);
-      }
-    }
-    tmpNotes.sort((a, b) => {
-      const dateA = a.date ? new Date(a.date).getTime() : 0;
-      const dateB = b.date ? new Date(b.date).getTime() : 0;
-      return dateA - dateB;
-    });
-    setNotes(tmpNotes);
-
-    console.debug("self notes updated");
-  };
-
-  useEffect(() => {
-    const selfNoteTopicsStr = localStorage.getItem(SELF_NOTE_TOPIC);
-    if (selfNoteTopicsStr) {
-      const tmpTopics = selfNoteTopicsStr.split(",");
-      setNoteRawTopics(tmpTopics.filter((t) => t.length === ADDRESS_HEX_LENGTH));
-    }
-  }, []);
-
-  useEffect(() => {
-    fetchNotes();
-  }, [noteRawTopics]);
-
-  useEffect(() => {
-    const privKey = localStorage.getItem("privKey");
-    const noRedirectedPaths = [ROUTES.WELCOME1, ROUTES.WELCOME2, ROUTES.WELCOME3, ROUTES.WELCOME4, ROUTES.TACONBOARDING, ROUTES.PROFILECREATION];
-    if (!privKey && !noRedirectedPaths.includes(location.pathname as ROUTES)) {
-      navigate(ROUTES.APP);
-    }
-  }, [navigate]);
+  useRouteProtection(location, navigate);
 
   return (
     <>
